News : Rename 'News' and expose Tags/Categories + archive pages.
A straw-person prototype making the 'News' section more blog-like, and a
rename to help suggest there may be other things than news in there.

Posts should now have a single `category` and multiple `tags`. They will
then be collected in the respective index pages.

As we're limited by GitHub Pages not supporting `jekyll-archives` we have
to roll our own. We need to include a stub page for each index. These can
be generated by the (new) `buildPostArchives.py` script. They need to be
committed to the repo so they'll be avialable to GitHub when it runs Jekyll.

Not done any web/css in a while so there may well be some rough edges.
Took the opportunity to re-jig the templates a little to make it slightly
easier to maintain.