Travis : Move html-proofer out of Gemfile to allow GitHub docker builds to run
GitHubPages only supports a white-listed subset of Gems. This causes builds
in environments that mirror what is available on GitHubPages to fail.

html-proofer is used on travis for validation of links, so this moves it t
a travis-only place.

We also remove the `bundle exec` around build/validate around html-proofer
as otherwise it only has the Gems in the `Gemfile` available. As we've just
built this environment via `bundle install` then there should be no
Adisparit.y