Add site-specific custom Sass stylings
- `gaff-variables.scss`: Global Sass variables for use in stylings,
plus overrides to Bootstrap variables and maps.
- `gaff-classes.scss`: General stylings, like a regular site.css, that
apply across pages. Of particular note are the styles under `main#main`
and `article#article`; these should standardize the apperance of
typography and images between main pages and posts.
- `gaff-code-blocks.scss`: Special stylings specific to code-blocks.
Isolated because they're based on Bootstrap's own website, and are
likely to change as Bootstrap versions change.