Update and add Liquid layouts
- `base.html`: The core HTML scaffolding that all pages use.
- `default.html`: For top-level content, such as pages included in the
nav (with the notable exception of the home page).
- `post.html`: Made to conform to Bootstrap 4. The banner at the top of
the post and thumbnail in the archive view are now the same `banner:`
image in the front matter, meaning authors won't need to awkwardly mix
the banner image into the content. Added includes for an
author, provision for post updates, and a transcluded list of the post's
 tags at the end of the content.