Update and clean up main site pages
- `index.md`: Updated to conform better to Bootstrap 4. Carousel shrunk
to work better with images of various sizes. Works much better
on mobile.
- `download.html`: Expanded content, simplified the presentation; made
it straight to the point, and covers developers' needs better.
- `blog.html`/`news.html`: Replaced with `news/index.html` to conform to
Jekyll blog archive requirements. Updated layout to be a little more
modern, and added a list of all blog tags.
- `tags.md`: Dynamic index of all tags. Stop-gap solution until Jekyll
provides for automatic page creation based on categories/tags.
- `community.html`/`forum.html`: Redirect straight to the Google group,
since putting a Google forum in an `<iframe>` causes display issues.
- Added a backup 404 page.
- Switched to HTTPS links where possible.