Update downloads page
Remove OSX link, because we don't have anyone to maintain the build/release process, and add Python 3 download in its place. The links won't work until we update the site Gaffer version to 0.59.0.0, but I want to do that separately to demonstrate the auto-update GitHub workflow I'm about to add.