GitHub Workflows : Update host OS to Ubuntu 20.04
As described at actions/virtual-environments#3287 support for Ubuntu 16.04 was dropped by GitHub on September 20th. Our CI workflows are now failing to start because they can't find a worker.

Our choices are 18.04 or 20.04. In the absence of any other reason, I've chosen 20.04 because it will last longer before we have to update again.