CI : Remove `htmlproofer --check-html` option
This feature was removed in `htmlproofer` 4.0. From the developer :

> This was intentionally removed as part of the html-proofer 4.0 major release. It was removed because, over the years, it has provided more false positives than actual errors. HTML 5 engines in particular are much more forgiving of broken HTML, and so it logging the four or five small errors this check actual looked for was not worth the trouble of maintainence (I'm just one developer). I know of no other tools which duplicate the feature but there might be some W3C validators which run a similar check.

We are much more interested in the link and image validation features, which continue to work.

Also remove `--empty-alt-ignore`, which I believe is now the default behaviour, based on https://github.com/gjtorikian/html-proofer/pull/717. And rename `--url-ignore` to `--ignore-urls`, because.