Replace Community with Resources and redirect
Rather than contain a no longer supported embed to the gaffer-dev group, the Community page has been replaced with a Resources page. The new Resources page links to various external gafferhq resources, some community maintained examples and extensions, as well as any community groups that we're aware of.

A redirect has been left from the original Community page to the new Resources page in the off-chance that anything was linking to it.